requirements.txt


colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1


dj-database-url==1.0.0
Django==4.1.2

django-filter==22.1

django-heroku==0.3.1
django-import-export==3.0.1

django-jazzmin==2.5.0
et-xmlfile==1.1.0
folium==0.13.0
fonttools==4.38.0
future==0.18.2
geocoder==1.38.1
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.6.1
numpy==1.23.4
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3


psycopg2==2.9.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0




tablib==3.2.1

whitenoise==6.2.0

